done with 104-fibonacci .c file

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -7,9 +7,9 @@
  */
 int main(void)
 {
-	unsigned long int fib1 = 1;
-	unsigned long int fib2 = 2;
-	unsigned long int next;
+	int fib1 = 1;
+	int fib2 = 2;
+	int next;
 	int n;
 
 	printf("%d, %d, ", fib1, fib2);
